add a link for "contact"
would it be possible to put a contact link on the main page along side "News" "Data" "Developer" "About"  Just a one stop hit for a quick contact.  I am having a hard time seeing where that line of links is located in code.  I added a dash to trigger a clickable button "propose changes"